backend/requirements.txt
django-celery-beat==2.8.1

# OCR関連
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
# その他
python-decouple==3.8

amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.14.2
billiard==4.2.2
bs4==0.0.2
celery==5.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.0.7
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
gunicorn==23.0.0
whitenoise==6.8.2











